Cleaned up CoroutineHost. Split up Coroutine's nested classes into two separete classes. Made TwitchUnityClient constructor ILogger default to null.
